+++ lib/_dev/_dev.dart
@@ -1,13 +1,11 @@
 import 'package:flutter/cupertino.dart';
 import 'package:flutter/material.dart';
-import 'package:twipe_flutter_core/_dev/_controller/main_page_controller.dart';
 import 'package:twipe_flutter_core/_dev/_pages/_main_page.dart';
 import 'package:twipe_flutter_core/twipe_flutter_core.dart';
 
 Future<void> main() async {
   await TwipeFlutterCore.initialize(
       environmentFilePath: 'resources/env.env.example');
-  MainPageController();
   runApp(DevApp());
 }