+++ lib/utils/cache/cache_object.dart
@@ -1,8 +1,10 @@
+import 'package:twipe_flutter_core/utils/json/json_object.dart';
+
 /// Cache Object class to Handle Data
-class CacheObject {
+class CacheObject extends JSONObject {
   final Map<String, dynamic> _data;
 
-  CacheObject(this._data);
+  CacheObject(this._data) : super(_data);
 
   /// CacheObject has Data
   bool hasData() {
@@ -12,4 +14,7 @@ class CacheObject {
   Map<String, dynamic> getData() {
     return _data;
   }
+
+  @override
+  List<String> get validator => [];
 }