+++ lib/widgets/page/watched/watched_page_state.dart
@@ -5,40 +5,9 @@ import 'watched_page.dart';
 
 /* WatchedPageState with watchable */
 abstract class WatchedPageState<T extends WatchedPage> extends State<T>
-    with Watchable, SingleTickerProviderStateMixin {
-  AnimationController? _loadingAnimationController;
-  Animation? _loadingAnimation;
-
-  int loadingAnimationDuration = 2;
-
-  double loadingAnimationBegin = 0.0;
-
-  double loadingAnimationEnd = 12.0;
-
-  bool loadingAnimationReverse = true;
-
-  AnimationController getLoadingAnimationController() {
-    return _loadingAnimationController!;
-  }
-
-  Animation getLoadingAnimation() {
-    return _loadingAnimation!;
-  }
-
-  bool useLoadableAnimation = true;
-
+    with Watchable {
   @override
   void initState() {
-    if (useLoadableAnimation) {
-      _loadingAnimationController = AnimationController(
-          vsync: this, duration: Duration(seconds: loadingAnimationDuration));
-      _loadingAnimation =
-          Tween(begin: loadingAnimationBegin, end: loadingAnimationEnd).animate(
-        CurvedAnimation(
-            parent: _loadingAnimationController!, curve: Curves.easeOut),
-      );
-      _loadingAnimationController!.repeat(reverse: loadingAnimationReverse);
-    }
     super.initState();
   }